+++ Nos/Models/Event+CoreDataClass.swift
@@ -67,7 +67,7 @@ extension FetchedResults where Element == Event {
 public class Event: NosManagedObject {
     
     static var replyNoteReferences = "kind = 1 AND ANY eventReferences.referencedEvent.identifier == %@ " +
-        "AND author.muted == false"
+        "AND author.muted = false"
     
     @nonobjc public class func allEventsRequest() -> NSFetchRequest<Event> {
         let fetchRequest = NSFetchRequest<Event>(entityName: "Event")
@@ -121,17 +121,19 @@ public class Event: NosManagedObject {
         let kind = EventKind.text.rawValue
         let featuredPredicate = NSPredicate(
             format: "kind = %i AND eventReferences.@count = 0 AND author.hexadecimalPublicKey IN %@ " +
-                "AND NOT author IN %@.follows.destination AND createdAt <= %@",
+                "AND NOT author IN %@.follows.destination AND NOT author = %@ AND createdAt <= %@ AND " +
+                "author.muted = false",
             kind,
             featuredAuthors.compactMap {
                 PublicKey(npub: $0)?.hex
             },
             currentUser,
+            currentUser,
             before as CVarArg
         )
-            
+        
         let twoHopsPredicate = NSPredicate(
-            format: "kind = %i AND eventReferences.@count = 0 " +
+            format: "kind = %i AND eventReferences.@count = 0 AND author.muted = false " +
                 "AND ANY author.followers.source IN %@.follows.destination AND NOT author IN %@.follows.destination",
             kind,
             currentUser,
@@ -774,7 +776,7 @@ public class Event: NosManagedObject {
     /// Returns the event this note is directly replying to, or nil if there isn't one.
     func referencedNote() -> Event? {
         if let rootReference = eventReferences?.first(where: {
-            ($0 as? EventReference)?.marker ?? "" == "reply"
+            ($0 as? EventReference)?.type == .reply
         }) as? EventReference,
             let referencedNote = rootReference.referencedEvent {
             return referencedNote
@@ -790,7 +792,7 @@ public class Event: NosManagedObject {
     /// Returns the root event of the thread that this note is replying to, or nil if there isn't one.
     func rootNote() -> Event? {
         let rootReference = eventReferences?.first(where: {
-            ($0 as? EventReference)?.marker ?? "" == "root"
+            ($0 as? EventReference)?.type == .root
         }) as? EventReference
         
         if let rootReference, let rootNote = rootReference.referencedEvent {